Technonews/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic[email]
pydantic-settings
sqlalchemy
aiosqlite
feedparser 